% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_comid_index}
\alias{add_comid_index}
\title{Add COMID index based on THREDDs file}
\usage{
add_comid_index(fileIDs)
}
\arguments{
\item{fileIDs}{a output from \code{fileID_find}}
}
\value{
data.frame
}
\description{
Add COMID index based on THREDDs file
}
\keyword{internal}
