% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_daily}
\alias{aggregate_daily}
\title{Aggregate Hourly Data to Daily Values}
\usage{
aggregate_daily(rawData, FUN = "mean")
}
\arguments{
\item{rawData}{data extracted with \code{readNWMdata}}

\item{FUN}{function used to aggregate values}
}
\value{
an aggregated data.frame
}
\description{
Aggregate Hourly Data to Daily Values
}
