% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_monthly}
\alias{aggregate_monthly}
\title{Aggregate Hourly Data to Monthly Values}
\usage{
aggregate_monthly(rawData, FUN = "sum")
}
\arguments{
\item{rawData}{data extracted with \code{readNWMdata}}

\item{FUN}{function used to aggregate values}
}
\value{
an aggregated data.frame
}
\description{
Aggregate Hourly Data to Monthly Values
}
