% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_season}
\alias{aggregate_season}
\title{Aggregate Hourly Data to Seasonal Values}
\usage{
aggregate_season(rawData, FUN = "mean", na.rm = TRUE)
}
\arguments{
\item{rawData}{data extracted with \code{readNWMdata}}

\item{FUN}{function used to aggregate values}
}
\value{
an aggregated data.frame
}
\description{
Aggregate Hourly Data to Seasonal Values
}
