% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggregate_yearly}
\alias{aggregate_yearly}
\title{Aggregate Hourly Data to Yearly Values}
\usage{
aggregate_yearly(rawData, FUN, na.rm = TRUE)
}
\arguments{
\item{rawData}{data extracted with \code{readNWMdata}}

\item{FUN}{function used to aggregate values}
}
\value{
an aggregated data.frame
}
\description{
Aggregate Hourly Data to Yearly Values
}
