% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cID}
\alias{cID}
\title{Index COMIDs
Dataset containing index information for NWM access}
\format{An object of class \code{array} of length 10000.}
\usage{
cID
}
\description{
Index COMIDs
Dataset containing index information for NWM access
}
\keyword{datasets}
