% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_comid.R
\name{discover_comid}
\alias{discover_comid}
\title{Discover NHD COMID}
\usage{
discover_comid(AOI, sf = FALSE)
}
\arguments{
\item{sf}{return the COMID (FALSE) or spatial catchment object (TRUE)}

\item{pt}{a sf point object}

\item{lat}{a numeric latitude}

\item{lon}{a numeric longitude}
}
\value{
a sf object or numeric COMID
}
\description{
Discover NHD COMID
}
