% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover_nhd}
\alias{discover_nhd}
\title{Discover NHD Features}
\usage{
discover_nhd(AOI, feature = NULL)
}
\arguments{
\item{AOI}{a sf point or polyfon object}

\item{feature}{if not NULL. What realization of an NHD object should be returned. Options include: catchment, flowline, and outlet.}
}
\value{
a sf object or numeric COMID
}
\description{
Discover NHD Features
}
