% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover_nwis}
\alias{discover_nwis}
\title{Discover USGS NWIS Stream Gages}
\usage{
discover_nwis(AOI = NULL)
}
\arguments{
\item{AOI}{A spatial geometry}
}
\value{
sf object
}
\description{
\code{discover_nwis} returns a POINT feature class of NWIS gages reporting parameter code '00060' (streamflow) for an Area of Interest that ALSO report NWM data. 
Data is accessed through the NWIS web portal.
}
