% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_retro.R
\name{extract_retro}
\alias{extract_retro}
\title{Extract Historic timeseries for COMID}
\usage{
extract_retro(comid, startDate = NULL, endDate = NULL,
  archive = NULL)
}
\arguments{
\item{comid}{an NHD COMID}

\item{startDate}{a start date (YYYY-MM-DD)}

\item{endDate}{an end date (YYYY-MM-DD)}

\item{archive}{path to the data archive}
}
\value{
data.frame
}
\description{
Extract Historic timeseries for COMID
}
