% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thredds.R
\name{extract_retro_url}
\alias{extract_retro_url}
\title{NWM Reanalysis Extraction}
\usage{
extract_retro_url(comid, startDate = NULL, endDate = NULL)
}
\arguments{
\item{comid}{a NHD common identifier}

\item{startDate}{a start date (YYYY-MM-DD)}

\item{endDate}{an end date (YYYY-MM-DD)}
}
\value{
data.frame
}
\description{
Download hourly flow values for an NHD COMID from the National Water Model version 2.0. Returned data is avaliable between "1993-01-01" and "2017-12-31" but can be subset using a startDate and endDate.
}
