% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileID_find}
\alias{fileID_find}
\title{Build COMID metadata}
\usage{
fileID_find(x)
}
\arguments{
\item{x}{an NHD COMID}
}
\value{
data.frame
}
\description{
Find the THREDDS file id and COMID index
}
\keyword{internal}
