% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{gage_lu}
\alias{gage_lu}
\title{COMID to NWIS ID lookup
Dataset containing COMID-NWIS lookup}
\format{
A data frame with 7541 rows and 2 variables
}
\usage{
gage_lu
}
\description{
COMID to NWIS ID lookup
Dataset containing COMID-NWIS lookup
}
\examples{
\dontrun{
 gage_lu = gage_lu
}
}
\keyword{datasets}
