% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hour_seq}
\alias{hour_seq}
\title{Time Sequence Generator}
\usage{
hour_seq(startDate, endDate)
}
\arguments{
\item{startDate}{}

\item{endDate}{}
}
\value{
vector of dates
}
\description{
Construct time series from start and end data
}
\keyword{internal}
