% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_index.R
\docType{data}
\name{nwm_retro_index}
\alias{nwm_retro_index}
\title{Index Files
Dataset containing index information for NWM access}
\format{An object of class \code{data.frame} with 273 rows and 3 columns.}
\usage{
nwm_retro_index
}
\description{
Index Files
Dataset containing index information for NWM access
}
\keyword{datasets}
