% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{readNWMdata}
\alias{readNWMdata}
\title{NWM Reanalysis Extraction}
\usage{
readNWMdata(
  comid = NULL,
  siteNumber = NULL,
  startDate = "1993-01-01",
  endDate = "2018-12-31",
  tz = "UTC",
  version = 2
)
}
\arguments{
\item{comid}{a NHD common identifier}

\item{siteNumber}{a USGS NWIS site number (eight digits)}

\item{startDate}{a start date (YYYY-MM-DD)}

\item{endDate}{an end date (YYYY-MM-DD)}

\item{tz}{the desired timezone of the data. Can be found with `OlsonNames()`}

\item{version}{the NWM version to extract (current = 1.2 or 2.0 (default))}
}
\value{
data.frame
}
\description{
Download hourly flow values for an NHD COMID from the National Water Model version 2.0. Returned data is availiable between "1993-01-01" and "2017-12-31" but can be subset using a startDate and endDate.
}
