% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retro_call}
\alias{retro_call}
\title{Retro Call Meta Data}
\usage{
retro_call(comid = NULL, startDate = NULL, endDate = NULL)
}
\arguments{
\item{comid}{an NHD COMID}

\item{startDate}{a start date (YYYY-MM-DD)}

\item{endDate}{an end date (YYYY-MM-DD)}
}
\value{
a data.frame
}
\description{
build out THREDDs call and meta information
}
\keyword{internal}
