% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aggregate.R
\name{split_time}
\alias{split_time}
\title{Split Y-M-D-H into time components}
\usage{
split_time(rawData, time_col)
}
\arguments{
\item{rawData}{rasData with time column}
}
\value{

}
\description{
Split Y-M-D-H into time components
}
