% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrract.R
\name{extract_retro}
\alias{extract_retro}
\title{Extract Historic timeseries for COMID}
\usage{
extract_retro(comid, archive = NULL)
}
\arguments{
\item{comid}{an NHD COMID}

\item{archive}{where the archive is located}
}
\value{
data.frame
}
\description{
Extract Historic timeseries for COMID
}
