% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{gage_lu}
\alias{gage_lu}
\title{COMID to NWIS ID lookup
Dataset containing index information for NWIS access}
\format{An object of class \code{data.frame} with 7541 rows and 2 columns.}
\usage{
gage_lu
}
\description{
COMID to NWIS ID lookup
Dataset containing index information for NWIS access
}
\keyword{datasets}
