% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_thredds.R
\name{readNWMdata}
\alias{readNWMdata}
\title{NWM Reanalysis Extraction}
\usage{
readNWMdata(comid = NULL, siteNumber = NULL, startDate = NULL,
  endDate = NULL)
}
\arguments{
\item{comid}{a NHD common identifier}

\item{startDate}{a start date (YYYY-MM-DD)}

\item{endDate}{an end date (YYYY-MM-DD)}
}
\value{
data.frame
}
\description{
Download hourly flow values for an NHD COMID from the National Water Model version 2.0. Returned data is availiable between "1993-01-01" and "2017-12-31" but can be subset using a startDate and endDate.
}
